+++ common/src/main/java/traben/entity_texture_features/mixin/MixinPackScreen.java
@@ -1,0 +1,48 @@
+package traben.entity_texture_features.mixin;
+
+import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.client.gui.screen.pack.PackScreen;
+import net.minecraft.client.gui.screen.pack.ResourcePackOrganizer;
+import net.minecraft.client.gui.widget.ButtonWidget;
+import net.minecraft.client.gui.widget.TexturedButtonWidget;
+import net.minecraft.resource.ResourcePackManager;
+import net.minecraft.screen.ScreenTexts;
+import net.minecraft.text.Text;
+import net.minecraft.util.Identifier;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import traben.entity_texture_features.ETFClientCommon;
+import traben.entity_texture_features.ETFVersionDifferenceHandler;
+import traben.entity_texture_features.config.screens.ETFConfigScreenMain;
+import traben.entity_texture_features.texture_handlers.ETFManager;
+import traben.entity_texture_features.utils.ETFUtils2;
+
+import java.io.File;
+import java.util.Objects;
+import java.util.function.Consumer;
+
+import static traben.entity_texture_features.ETFClientCommon.ETFConfigData;
+import static traben.entity_texture_features.ETFClientCommon.MOD_ID;
+
+@Mixin(PackScreen.class)
+public abstract class MixinPackScreen extends Screen {
+
+
+    protected MixinPackScreen(Text title) {
+        super(title);
+    }
+
+    @Inject(method = "init", at = @At("TAIL"))
+    private void etf$illegalPathOverride(CallbackInfo ci) {
+        this.addDrawableChild(new TexturedButtonWidget((this.width-48), (this.height-40), 32, 32,
+                0,0,32, new Identifier(MOD_ID+":textures/gui/settings.png"),32,64,
+                (button) -> {
+                    Objects.requireNonNull(client).setScreen(new ETFConfigScreenMain(this));
+                }));
+    }
+}
+
+