+++ CHANGE_LOG.md
@@ -1,5 +1,10 @@
 [**ETF Changelog:**
 
+[V4.1.dev.2]
+- tweaked new property names
+- added another property distanceToPlayer
+- improved skin tool added enchanted and emissive options
+
 [V4.1.dev.1]
 - cloth config no longer required on forge or included in fabric
 - updated russian translation thanks to @Felix14-v2
@@ -7,8 +12,7 @@
 - fixed a crash when reading corrupt or broken config files
 - complete GUI overhaul
 - added GUI for selecting player features and printing a player feature ready skin with all your choices *(W.I.P)*
-- added many new texture properties *(speedRange, jumpRange, maxHealth, llamaInventory, trapHorse, angry, 
-wardenAngerLevel, angryAtClient, hiddenGene, isPlayerCreated, isScreamingGoat)*
+- added many new texture properties
 
 [V4.0.2]
 - fixed ArrayIndexOutOfBoundsException crash